requirements.txt
# Flask backend dependencies
flask>=2.0.0
flask-cors>=4.0.0
requests>=2.28.0
python-dotenv>=1.0.0
Pillow>=9.0.0

# Image processing dependencies
numpy>=1.21.0
opencv-python>=4.5.0  # Optional: for additional image processing capabilities

# Development tools
pytest>=7.0.0  # For running tests
black>=22.0.0  # For code formatting
flake8>=4.0.0  # For linting 